remove CRT dependency from HTTPAPI
